+++ app/src/main/java/de/ur/explure/views/CategorieOneFragment.kt
@@ -14,10 +14,10 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [TestFragment.newInstance] factory method to
+ * Use the [CategorieOneFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class TestFragment : Fragment() {
+class CategorieOneFragment : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
@@ -33,7 +33,7 @@ class TestFragment : Fragment() {
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
                               savedInstanceState: Bundle?): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_test, container, false)
+        return inflater.inflate(R.layout.fragment_categorie_one, container, false)
     }
 
     companion object {
@@ -43,12 +43,12 @@ class TestFragment : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment TestFragment.
+         * @return A new instance of fragment CategorieOneFragment.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-                TestFragment().apply {
+                CategorieOneFragment().apply {
                     arguments = Bundle().apply {
                         putString(ARG_PARAM1, param1)
                         putString(ARG_PARAM2, param2)
