--- app/src/main/res/layout/fragment_test.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
-    tools:context=".views.TestFragment">
-
-    <ImageView
-        android:id="@+id/imageView"
-        android:layout_width="507dp"
-        android:layout_height="869dp"
-        tools:src="@tools:sample/avatars" />
-</FrameLayout>
