+++ app/src/main/res/navigation/nav_graph.xml
@@ -12,17 +12,8 @@
         tools:layout="@layout/fragment_home"/>
     <fragment
         android:id="@+id/search"
-        android:name="de.ur.explure.views.Search"
+        android:name="de.ur.explure.views.SearchFragment"
         android:label="fragment_search"
-        tools:layout="@layout/fragment_search" >
-        <action
-            android:id="@+id/action_search_to_testFragment"
-            app:destination="@id/testFragment" />
-    </fragment>
-    <fragment
-        android:id="@+id/testFragment"
-        android:name="de.ur.explure.views.TestFragment"
-        android:label="fragment_test"
-        tools:layout="@layout/fragment_test" />
+        tools:layout="@layout/fragment_search" />
 
 </navigation>
