+++ app/src/main/res/values/strings.xml
@@ -8,6 +8,8 @@
     <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
 
+    <string name="lorem_ipsum">Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua.  </string>
+
 
     <string name="search_button">Suche</string>
     <string name="kategorie_text">Übersicht Kategorien</string>
